backend/requirements.txt
Flask
flask-cors
flask-sqlalchemy
gunicorn
gTTS
SpeechRecognition
Werkzeug
PyJWT